Add ContactDaoTest with first test and setup
In setUp we assign test database and fire up sessionFactory by initializing ContactDaoImpl.

In @After method we close database.

The first test uses AssertJ and checks that contact can be saved to empty database.

And actually findOne method as well...so yeah not so good Unit Test I guess ... whatever